CI-60 Refactor the new attraction away from the updated attraction
- Replicates the add attraction as an update attraction.
- Adds a layer group for holding the markers (and later removing them).
- Beefs up the understanding of waiting for the *view* to settle before
we can create the map instance; kicks off the Leaflet Map implementation
after the containing page is about to enter the View (meaning all components
are assembled and scoped out for size, maybe?).

Performs a bit more of the conversion from Location to Attraction.

This work also identified other spots that need attention and ties together
some of the outstanding problems we've faced with the Map life-cycle.