CI-102 Plop down the Team Assembled button
- Brings in TeamPage code and HTML from Ionic 3 app including the button.
- Patches in the spec code.

This also updates the show-game spec to not bother with defining the provider for HttpClient;
that code would only be brought in for services that talk to the back end.

If a spec asks for a provider of HttpClient, it could be because the spec
has gotten ahold of an actual service instead of a mocked service.