CI-123 Review the Outing page
- Adds Pinned Map component.
- Brings Outing Page up to Ionic 4.

Timing for display of the map remains tricky. This implementation asks
the PinnedMap to wait on drawing the map until it gets an Attraction
from the encompassing parent page. The component doesn't respond to
`ionViewDidEnter`, but the page does. However, we only want to initialize
the map once, so it becomes static and endures for the life of the app.