CI-254 Instead of showing the Start Location, we should show the Outing View when first coming up
- Specifically routes to the Outing page when the game hasn't yet started.
- When returning to the outing page from viewing the Clue Ride website, the attraction doesn't get loaded and
an exception was being thrown. This led to a couple of things:
  - Removed the unused 'Pin' LatLon value for the pinned-map.component.ts.
  - Added a Guard to the Outing page to wait until the data load had completed.

Waiting for DataLoad might be a good thing to add to a bunch of guards.
Instead of writing the Guard such that it is page-specific, we can make it condition specific -- the inclusion of
the guard states what kind of wait is being performed.